proj1_AI/requirements.txt
openai
requests
python-dotenv
azure-cognitiveservices-speech